Crate miniupnpc

MiniUPnP

This crate contains safe bindings to the miniupnpc library.

"UPnP is a set of networking protocols that allows networked devices to seamlessy discover each other's presence on the network."

See also: Universal Plug and Play

Modules

commands
UPnP Commands

Structs

Device
Single uPnP device.

DeviceList
uPnP discovered devices list

DeviceListIter
Iterator over a DeviceList

Error
MiniUPnP library errors.

IgdData
IGD data

IgdService
IGD service information

Urls
UPnP IGD device URLs

ValidIgd
Information about a valid IGD

Enums

LocalPort

Functions

api_version
MiniUPnP API version

discover
Discover uPnP devices.

version
MiniUPnP version